+++ src/main/java/pl/asie/computronics/integration/buildcraft/StatementParameters.java
@@ -1,0 +1,27 @@
+package pl.asie.computronics.integration.buildcraft;
+
+import buildcraft.api.statements.IStatementParameter;
+import buildcraft.api.statements.StatementManager;
+import pl.asie.computronics.integration.buildcraft.parameters.ActionParameterLampColor;
+
+/**
+ * @author Vexatos
+ */
+public enum StatementParameters {
+	Lamp_Color("lamp_color", ActionParameterLampColor.class);
+
+	public static final StatementParameters[] VALUES = values();
+	public final String name;
+	private Class<? extends IStatementParameter> paramClass;
+
+	private StatementParameters(String name, Class<? extends IStatementParameter> paramClass) {
+		this.name = name;
+		this.paramClass = paramClass;
+	}
+
+	public static void initialize() {
+		for(StatementParameters param : VALUES) {
+			StatementManager.registerParameterClass(param.paramClass);
+		}
+	}
+}