+++ src/main/java/pl/asie/computronics/integration/waila/ConfigValues.java
@@ -9,11 +9,11 @@ import pl.asie.computronics.reference.Mods;
  */
 enum ConfigValues {
 
-	Address(Mods.Computronics + ".enableAddress"),
+	//Address(Mods.Computronics + ".enableAddress"),
 	Tape(Mods.Computronics + ".enableTape"),
 	TapeName(Mods.Computronics + ".enableTapeName"),
-	DriveState(Mods.Computronics + ".enableDriveState"),
-	RelayBound(Mods.Computronics + ".enableRelayBound");
+	DriveState(Mods.Computronics + ".enableDriveState");
+	//RelayBound(Mods.Computronics + ".enableRelayBound");
 
 	private String key;
 	private boolean defvalue;