+++ src/main/java/pl/asie/computronics/integration/waila/WailaLocomotiveRelay.java
@@ -4,7 +4,7 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.util.StatCollector;
+import pl.asie.computronics.util.StringUtil;
 
 import java.util.List;
 
@@ -17,13 +17,13 @@ public class WailaLocomotiveRelay extends WailaComputronics {
 	public List<String> getWailaBody(ItemStack itemStack, List<String> currenttip, IWailaDataAccessor accessor,
 		IWailaConfigHandler config) {
 
-		if(!ConfigValues.RelayBound.getConfig(config)) {
-			return currenttip;
-		}
+		//if(!ConfigValues.RelayBound.getConfig(config)) {
+		//	return currenttip;
+		//}
 
 		NBTTagCompound nbt = accessor.getNBTData();
 		String boundKey = "tooltip.computronics.waila.relay." + (nbt.getBoolean("bound") ? "bound" : "notbound");
-		currenttip.add(StatCollector.translateToLocal(boundKey));
+		currenttip.add(StringUtil.localize(boundKey));
 		return currenttip;
 	}
 }