+++ src/main/java/pl/asie/computronics/integration/waila/WailaTapeDrive.java
@@ -4,9 +4,10 @@ import mcp.mobius.waila.api.IWailaConfigHandler;
 import mcp.mobius.waila.api.IWailaDataAccessor;
 import mcp.mobius.waila.api.SpecialChars;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.StatCollector;
 import pl.asie.computronics.Computronics;
+import pl.asie.computronics.api.tape.IItemTapeStorage;
 import pl.asie.computronics.tile.TileTapeDrive;
+import pl.asie.computronics.util.StringUtil;
 
 import java.util.List;
 import java.util.Locale;
@@ -23,21 +24,21 @@ public class WailaTapeDrive extends WailaComputronics {
 
 		TileTapeDrive drive = (TileTapeDrive) accessor.getTileEntity();
 		ItemStack is = drive.getStackInSlot(0);
-		if(is != null && is.getItem() == Computronics.itemTape) {
+		if(is != null && is.getItem() instanceof IItemTapeStorage) {
 			String label = Computronics.itemTape.getLabel(is);
 			if(label.length() > 0 && ConfigValues.TapeName.getConfig(config)) {
-				currenttip.add(StatCollector.translateToLocalFormatted("tooltip.computronics.waila.tape.labeltapeinserted",
+				currenttip.add(StringUtil.localizeAndFormat("tooltip.computronics.waila.tape.labeltapeinserted",
 					label + SpecialChars.RESET));
 			} else {
-				currenttip.add(StatCollector.translateToLocal("tooltip.computronics.waila.tape.tapeinserted"));
+				currenttip.add(StringUtil.localize("tooltip.computronics.waila.tape.tapeinserted"));
 			}
 			if(ConfigValues.DriveState.getConfig(config)) {
-				currenttip.add(StatCollector.translateToLocalFormatted("tooltip.computronics.waila.tape.state",
-					StatCollector.translateToLocal("tooltip.computronics.waila.tape.state."
+				currenttip.add(StringUtil.localizeAndFormat("tooltip.computronics.waila.tape.state",
+					StringUtil.localize("tooltip.computronics.waila.tape.state."
 						+ drive.getEnumState().toString().toLowerCase(Locale.ENGLISH))));
 			}
 		} else {
-			currenttip.add(StatCollector.translateToLocal("tooltip.computronics.waila.tape.notapeinserted"));
+			currenttip.add(StringUtil.localize("tooltip.computronics.waila.tape.notapeinserted"));
 		}
 		return currenttip;
 	}